code-interpreter-demo/pyproject.toml
[project]
name = "code-interpreter-demo"
version = "0.1.0"
description = "AgentRun Code Interpreter Sandbox 代码执行示例"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agentrun-sdk[playwright,server]>=0.0.8",
    "python-dotenv>=1.0.0",
]
